pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "supervisor-distillation"
version = "0.1.0"
description = "Supervisor-worker deep research via programmatic distillation"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "python-dotenv",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
